+++ test/setup.js
@@ -1,2 +1,3 @@
+process.env.SWCRC = true
 process.env.NODE_ENV = 'test'
 global.IS_REACT_ACT_ENVIRONMENT = true